# performance improvements for scripts/update_all_products.pl