--- a/.gitignore
+++ b/.gitignore
@@ -17,6 +17,7 @@ Thumbs.db
 Store.debug.txt
 *.log
 logs/modperl_error_log
+nytprof*.out
 
 # Local databases
 data/mongodb