Some low-hanging-fruit in avoiding looping over CORE:regcomp. It feels like there's more to squeeze out, but I don't know the data structures well enough. It's also hard to compare like with like, if you let the code actually update the stored data.

### Possible improvements
### 1)
A section that still has slow regexes looped over a lot is the
```
($new_ingredient =~ /(^($regexp)\b|\b($regexp)$)/i)
($new_ingredient =~ /($regexp)$/i) )
($new_ingredient =~ /(^($regexp)|($regexp)$)/i)
```
expressions in: 
https://github.com/svensven/openfoodfacts-server/blob/01235cb7d8e77b7d62e44efba8528898e828ba19/lib/ProductOpener/Ingredients.pm#L1245-L1263

### 2)
Also, preparse_ingredients_text() is called twice, presumably with the same inputs:
```
# spent 8.49s (914ms+7.57) within ProductOpener::Ingredients::preparse_ingredients_text which was called 490 times, avg 17.3ms/call: 
# 245 times (519ms+4.15s) by ProductOpener::Ingredients::parse_ingredients_text at line 823, avg 19.1ms/call 
# 245 times (395ms+3.42s) by ProductOpener::Ingredients::extract_ingredients_classes_from_text at line 3984, avg 15.6ms/call
```
from update_all_products.pl:
```
			extract_ingredients_from_text($product_ref);
			extract_ingredients_classes_from_text($product_ref);
```
I thought about trying a static/state variable, but it would have to be done in a way that didn't just keep consuming more RAM.

### Pictures
example before & after:

![before](https://user-images.githubusercontent.com/5736616/93671222-1a83b180-fa99-11ea-8268-5b3e1653f419.png)
![after](https://user-images.githubusercontent.com/5736616/93671253-5b7bc600-fa99-11ea-85f6-f2dcea731ffc.png)