@stephanegigandet 2020-09-19
Looks very good to me.

Regarding preparse_ingredients_text() being called twice, it definitely does not make sense, as the 2 functions that call it are always called one after the other anyway.

One way could be to change the function to pass $product_ref to it, so that we can store the results in $product_ref->{ingredients_text_preparsed}.
Then we would need to remove it after calling the 2 functions, so that we don't store it in mongodb and the .sto files.